show WorkspaceSwitcherPopup
The idea is taken from micheleg's dash-to-dock extension, which can also
switch workspaces when scrolled over the dock, but shows the indicator
when doing so. This tells the current position more clearly which users
may appreciate.

If both extensions are installed, they even share the switcher applet
object without duplication.